+++ src/Symfony/Component/Validator/Constraints/Choice.php
@@ -32,11 +32,6 @@ class Choice extends Constraint
         self::TOO_MANY_ERROR => 'TOO_MANY_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $choices;
     public $callback;
     public $multiple = false;