+++ src/Symfony/Component/Validator/Constraints/Country.php
@@ -30,11 +30,6 @@ class Country extends Constraint
         self::NO_SUCH_COUNTRY_ERROR => 'NO_SUCH_COUNTRY_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $message = 'This value is not a valid country.';
     public $alpha3 = false;