+++ src/Symfony/Component/Validator/Constraints/Currency.php
@@ -31,11 +31,6 @@ class Currency extends Constraint
         self::NO_SUCH_CURRENCY_ERROR => 'NO_SUCH_CURRENCY_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $message = 'This value is not a valid currency.';
 
     public function __construct(array $options = null, string $message = null, array $groups = null, mixed $payload = null)