+++ src/Symfony/Component/Validator/Constraints/Date.php
@@ -30,11 +30,6 @@ class Date extends Constraint
         self::INVALID_DATE_ERROR => 'INVALID_DATE_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $message = 'This value is not a valid date.';
 
     public function __construct(array $options = null, string $message = null, array $groups = null, mixed $payload = null)