+++ src/Symfony/Component/Validator/Constraints/Email.php
@@ -28,10 +28,6 @@ class Email extends Constraint
     public const VALIDATION_MODE_HTML5_ALLOW_NO_TLD = 'html5-allow-no-tld';
     public const VALIDATION_MODE_HTML5 = 'html5';
     public const VALIDATION_MODE_STRICT = 'strict';
-    /**
-     * @deprecated since Symfony 6.2, use VALIDATION_MODE_HTML5 instead
-     */
-    public const VALIDATION_MODE_LOOSE = 'loose';
 
     public const INVALID_FORMAT_ERROR = 'bd79c0ab-ddba-46cc-a703-a7a4b08de310';
 
@@ -39,17 +35,11 @@ class Email extends Constraint
         self::VALIDATION_MODE_HTML5_ALLOW_NO_TLD,
         self::VALIDATION_MODE_HTML5,
         self::VALIDATION_MODE_STRICT,
-        self::VALIDATION_MODE_LOOSE,
     ];
 
     protected const ERROR_NAMES = [
         self::INVALID_FORMAT_ERROR => 'STRICT_CHECK_FAILED_ERROR',
     ];
-
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
 
     public $message = 'This value is not a valid email address.';
     public $mode;
@@ -72,10 +62,6 @@ class Email extends Constraint
         $this->message = $message ?? $this->message;
         $this->mode = $mode ?? $this->mode;
         $this->normalizer = $normalizer ?? $this->normalizer;
-
-        if (self::VALIDATION_MODE_LOOSE === $this->mode) {
-            trigger_deprecation('symfony/validator', '6.2', 'The "%s" mode is deprecated. It will be removed in 7.0 and the default mode will be changed to "%s".', self::VALIDATION_MODE_LOOSE, self::VALIDATION_MODE_HTML5);
-        }
 
         if (self::VALIDATION_MODE_STRICT === $this->mode && !class_exists(StrictEmailValidator::class)) {
             throw new LogicException(sprintf('The "egulias/email-validator" component is required to use the "%s" constraint in strict mode. Try running "composer require egulias/email-validator".', __CLASS__));