+++ src/Symfony/Component/Validator/Constraints/Image.php
@@ -58,11 +58,6 @@ class Image extends File
         self::CORRUPTED_IMAGE_ERROR => 'CORRUPTED_IMAGE_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $mimeTypes = 'image/*';
     public $minWidth;
     public $maxWidth;