+++ src/Symfony/Component/Validator/Constraints/Ip.php
@@ -70,16 +70,6 @@ class Ip extends Constraint
         self::INVALID_IP_ERROR => 'INVALID_IP_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const VERSIONS instead
-     */
-    protected static $versions = self::VERSIONS;
-
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $version = self::V4;
 
     public $message = 'This is not a valid IP address.';
@@ -100,8 +90,8 @@ class Ip extends Constraint
         $this->message = $message ?? $this->message;
         $this->normalizer = $normalizer ?? $this->normalizer;
 
-        if (!\in_array($this->version, self::$versions)) {
-            throw new ConstraintDefinitionException(sprintf('The option "version" must be one of "%s".', implode('", "', self::$versions)));
+        if (!\in_array($this->version, static::VERSIONS, true)) {
+            throw new ConstraintDefinitionException(sprintf('The option "version" must be one of "%s".', implode('", "', static::VERSIONS)));
         }
 
         if (null !== $this->normalizer && !\is_callable($this->normalizer)) {