+++ src/Symfony/Component/Validator/Constraints/Json.php
@@ -28,11 +28,6 @@ class Json extends Constraint
         self::INVALID_JSON_ERROR => 'INVALID_JSON_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $message = 'This value should be valid JSON.';
 
     public function __construct(array $options = null, string $message = null, array $groups = null, mixed $payload = null)