+++ src/Symfony/Component/Validator/Constraints/Regex.php
@@ -29,11 +29,6 @@ class Regex extends Constraint
         self::REGEX_FAILED_ERROR => 'REGEX_FAILED_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public $message = 'This value is not valid.';
     public $pattern;
     public $htmlPattern;