+++ src/Symfony/Component/Validator/Constraints/Timezone.php
@@ -41,11 +41,6 @@ class Timezone extends Constraint
         self::TIMEZONE_IDENTIFIER_INTL_ERROR => 'TIMEZONE_IDENTIFIER_INTL_ERROR',
     ];
 
-    /**
-     * @deprecated since Symfony 6.1, use const ERROR_NAMES instead
-     */
-    protected static $errorNames = self::ERROR_NAMES;
-
     public function __construct(
         int|array $zone = null,
         string $message = null,